+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,8 +1,20 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
-## Version 1.18.0
+## Version 1.19.0
 
 - Delete `ForwardedByActorNumber` and `ForwardedByActorRole` from `ForwardMeteredDataRejectedV1`.
+
+## Version 1.18.1
+
+- Use `IEnqueueAcceptedDataDto`, `IEnqueueRejectedDataDto` and `IEnqueueDataDto` marker interfaces.
+
+## Version 1.18.0
+
+- Add custom query `CalculationByIdQueryV1`.
+
+## Version 1.17.2
+
+- No functional changes.
 
 ## Version 1.17.1